[Feat] 프로필 삭제 시 요청도 삭제되도록 구현
- 프로필 삭제 시 DB에서 해당 유저가 발신자, 수신자로 설정된 요청이 모두 삭제되도록 구현

+++ src/database/firebase.ts
@@ -298,9 +298,30 @@ export const updateProfile = async (
   }
 };
 
-export const deleteProfile = async (id: string, tab: string) => {
+export const deleteProfile = async (
+  id: string,
+  tab: string,
+  userId: string | undefined
+) => {
+  console.log(userId);
+  if (!userId) return;
   const changedTabName = changeTabName(tab);
   await deleteDoc(doc(db, changedTabName, id));
+  await deleteRequest("recipientUserId", userId);
+  await deleteRequest("senderUserId", userId);
+};
+
+const deleteRequest = async (fieldName: string, userId: string) => {
+  const requestQuery = query(
+    collection(db, "requests"),
+    where(fieldName, "==", userId)
+  );
+  const snapshot = await getDocs(requestQuery);
+  if (snapshot.empty) {
+    return;
+  } else {
+    return await deleteDoc(doc(db, "requests", snapshot.docs[0].id));
+  }
 };
 
 export const getProfilesFromRequests = async (
